#54 Separate schema building from actual schema map, allow modification of schema using events
