+++ Civi/API/Service/Schema/SchemaMap.php
@@ -3,9 +3,6 @@
 namespace Civi\API\Service\Schema;
 
 use Civi\API\Service\Schema\Joinable\Joinable;
-use Civi\API\Service\Schema\Joinable\OptionValueJoinable;
-use CRM_Core_DAO_AllCoreTables as TableHelper;
-use CRM_Utils_Array as ArrayHelper;
 
 class SchemaMap {
 
@@ -15,89 +12,6 @@ class SchemaMap {
    * @var Table[]
    */
   protected $tables = array();
-
-  /**
-   * @param bool $autoload
-   *   Whether to autoload the tables using core classes. If not tables must
-   *   be added manually
-   */
-  public function __construct($autoload = true) {
-    if ($autoload) {
-      $this->loadTables();
-    }
-  }
-
-  /**
-   * Add all tables and joins
-   */
-  private function loadTables() {
-    /** @var \CRM_Core_DAO $daoName */
-    foreach (TableHelper::get() as $daoName => $data) {
-      $table = new Table($data['table']);
-      foreach ($daoName::fields() as $field => $fieldData) {
-        $this->addJoins($table, $field, $fieldData);
-      }
-      $this->addTable($table);
-    }
-
-    // add back references
-    foreach ($this->tables as $table) {
-      foreach ($table->getTableLinks() as $link) {
-
-        // there are too many possible joins from option value so skip
-        if ($link instanceof OptionValueJoinable) {
-          continue;
-        }
-
-        $target = $this->getTableByName($link->getTargetTable());
-        $joinable = new Joinable($link->getBaseTable(), $link->getBaseColumn());
-        $target->addTableLink($link->getTargetColumn(), $joinable);
-      }
-    }
-  }
-
-  /**
-   * @param Table $table
-   * @param string $field
-   * @param array $data
-   */
-  private function addJoins(Table $table, $field, array $data) {
-    $pseudoConstant = ArrayHelper::value('pseudoconstant', $data);
-    $fkClass = ArrayHelper::value('FKClassName', $data);
-
-    // can there be multiple methods e.g. pseudoconstant and fkclass
-    if ($fkClass) {
-      $tableName = TableHelper::getTableForClass($fkClass);
-      $fkKey = ArrayHelper::value('FKKeyColumn', $data, 'id');
-      $table->addTableLink($field, new Joinable($tableName, $fkKey));
-    } else if ($pseudoConstant) {
-      $this->addPseudoConstantJoin($table, $field, $pseudoConstant);
-    }
-  }
-
-  /**
-   * @param Table $table
-   * @param string $field
-   * @param array $data
-   */
-  private function addPseudoConstantJoin(Table $table, $field, array $data) {
-    $tableName = ArrayHelper::value('table', $data);
-    $optionGroupName = ArrayHelper::value('optionGroupName', $data);
-
-    if ($tableName) {
-      $keyColumn = ArrayHelper::value('keyColumn', $data, 'id');
-      $alias = str_replace('civicrm_', '', $tableName);
-      $joinable = new Joinable($tableName, $keyColumn, $alias);
-      $condition = ArrayHelper::value('condition', $data);
-      if ($condition) {
-        $joinable->addCondition($condition);
-      }
-      $table->addTableLink($field, $joinable);
-    } elseif ($optionGroupName) {
-      $joinable = new OptionValueJoinable($optionGroupName);
-      $table->addTableLink($field, $joinable);
-    }
-  }
 
   /**
    * @param $baseTableName
