+++ services.xml
@@ -6,7 +6,12 @@
 
         <service id="spec_gatherer" class="Civi\API\Spec\SpecGatherer"/>
 
-        <service id="schema_map" class="Civi\API\Service\Schema\SchemaMap"/>
+        <service id="schema_map_builder" class="Civi\API\Service\Schema\SchemaMapBuilder" public="false">
+            <argument type="service" id="dispatcher" />
+        </service>
+
+        <!-- factory service is deprecated but for now we're stuck with v2.5 -->
+        <service id="schema_map" class="Civi\API\Service\Schema\SchemaMap" factory-service="schema_map_builder" factory-method="build"/>
 
         <service id="joiner" class="Civi\API\Service\Schema\Joiner">
             <argument type="service" id="schema_map"/>
